# contracts/saas/openapi.yaml
openapi: 3.1.0
...
info:
  title: Dockling hosted API contract
  version: 0.4.1
  description: Versioned SaaS design. Hosted Phases 1–4 are implemented for synthetic
    local acceptance; billing, public deployment and future aliases remain planned.
    Uploads/workbook apply use bounded raw bytes. AI uses separate consent and conservative
    reservations; production teardown recovery remains a deployment gate.
servers:
- url: /api/v1
security:
- SessionCookie: []
paths:
...
  /connections:
    get:
      summary: Connections
      operationId: hosted_get__connections
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      - viewer
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
    post:
      summary: Create
      operationId: hosted_post__connections
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedConnectionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
      parameters:
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
  /connections/{connection}:
    put:
      operationId: put_connections_connection
      summary: PUT /connections/{connection}
      x-phase: 4
      x-feature-ids:
      - F24
      - F26
      - F27
      x-access-roles:
      - owner
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathConnection'
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionUpdate'
      x-implementation: planned; see implemented hosted connection routes
    delete:
      operationId: delete_connections_connection
      summary: DELETE /connections/{connection}
      x-phase: 4
      x-feature-ids:
      - F24
      x-access-roles:
      - owner
      x-workspace-scoped: true
      responses:
        '204':
          description: Completed
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathConnection'
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCommand'
      x-implementation: planned; see implemented hosted connection routes
    get:
      summary: Connection
      operationId: hosted_get__connections_connection
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      - viewer
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
...
  /admin/profiles/{profile}/tests:
    post:
      summary: Profile Test
      operationId: hosted_post__admin_profiles_profile_tests
      parameters:
      - name: profile
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedProfileTest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Profile Test Api V1 Admin Profiles  Identifier  Tests
                  Post
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - platform_admin
      x-feature-ids:
      - F29
      - F32
...
  /connections/{connection}/key:
    put:
      summary: Replace Key
      operationId: hosted_put__connections_connection_key
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedKeyCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
  /connections/{connection}/revoke:
    post:
      summary: Revoke
      operationId: hosted_post__connections_connection_revoke
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedRevisionCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
  /connections/{connection}/test:
    post:
      summary: Models
      operationId: hosted_post__connections_connection_test
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedRevisionCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
  /connections/{connection}/models:
    post:
      summary: Models
      operationId: hosted_post__connections_connection_models
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedRevisionCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
  /connections/{connection}/models/manual:
    post:
      summary: Manual
      operationId: hosted_post__connections_connection_models_manual
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedManualModel'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
  /connections/{connection}/selection:
    post:
      summary: Selection
      operationId: hosted_post__connections_connection_selection
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedSelection'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
  /connections/{connection}/terms:
    post:
      summary: Terms
      operationId: hosted_post__connections_connection_terms
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedTerms'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: dockling_session
      description: Opaque server session; Secure, HttpOnly, SameSite. Not an OIDC
        access token or browser role.
    PaymentSignature:
      type: apiKey
      in: header
      name: Payment-Signature
      description: Placeholder for the selected provider signature scheme; validate
        raw provider bytes, timestamp and replay policy before processing.
  schemas:
...
    HostedConnectionList:
      type: object
      additionalProperties: false
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/HostedConnectionRead'
        next_cursor:
          type: 'null'
...
  responses:
    Error400:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Nonrevealing missing or unauthorized private resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: Revision/idempotency/state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Error422:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error429:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error503:
      description: Dependency unavailable; no demo fallback
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    HeaderXCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        maxLength: 150
...
    PathConnection:
      name: connection
      in: path
      required: true
      schema: *id001
...